backend/requirements.txt
azure-functions
azure.cosmos
flake8
pytest
